requirements.txt
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1 
torch==1.7.0
torchvision==0.8.1
matplotlib
seaborn
pandas
tqdm
scikit-image
filterpy
schedule
configparser>=4.0.2
